Fixed bugs on the map screen, added pop-up for Set Start button, and removed clutter
If user clicks Start Nav on a marker, then clicked a new marker, they would immediately be navigated to that location.
If user navigates to a professor's office, then clicks a different marker before clicking Stop, it will zoom in on the professor's office marker, not the other marker.
If user skipped the tutorial, the button stays highlighted.